实现图书添加功能

+++ app/src/main/java/cn/eviao/bookstorage/model/Tag.kt
@@ -6,9 +6,10 @@ import androidx.room.PrimaryKey
 
 @Entity(
     tableName = "tags",
-    indices = [Index(value = ["text"], unique = true)])
+    indices = [Index(value = ["text"], unique = true)]
+)
 data class Tag(
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
     val text: String
 )
